SmartIX/requirements.txt
psycopg2-binary
matplotlib
numpy
pandas